task1/requirements.txt
numpy
torch
tqdm
faiss-gpu
transformers
pandas
PIL
sentence_transformers
clip_video_encode
